Moves the addon service provider into the Addon module.
